快要上线啦！

+++ src/main/java/com/forte/demo/listener/PrayListener.java
@@ -4,6 +4,7 @@ import com.forte.demo.bean.Person;
 import com.forte.demo.service.PersonService;
 import com.forte.demo.service.PrayService;
 import com.forte.demo.utils.PrayEnum;
+import com.forte.demo.utils.PrayUtils;
 import com.forte.qqrobot.anno.Filter;
 import com.forte.qqrobot.anno.Listen;
 import com.forte.qqrobot.anno.depend.Beans;
@@ -39,10 +40,15 @@ public class PrayListener {
     @Filter(value = "公主单抽")
     @Listen(MsgGetTypes.groupMsg)
     public void highOne(GroupMsg msg, MsgSender sender) throws Exception {
-        personService.reduceStar(new Person(msg.getQQ(),ONE,null));
-        String high = prayService.highOne(msg.getQQ());
-        CQCode cqCode_at = CQCodeUtil.build().getCQCode_At(msg.getQQ());
-        Integer baodiNum = prayService.getBaodiNum(PrayEnum.high, msg.getQQ());
+        String qq = msg.getQQ();
+        CQCode cqCode_at = CQCodeUtil.build().getCQCode_At(qq);
+        if (personService.getStar(qq)<ONE){
+            sender.SENDER.sendGroupMsg(msg.getGroup(),cqCode_at+" 积分不足！");
+            return;
+        }
+        personService.reduceStar(new Person(qq,ONE,null,null));
+        String high = prayService.highOne(qq);
+        Integer baodiNum = prayService.getBaodiNum(PrayEnum.high, qq);
         String baodila = "";
         if (baodiNum<=BAODI){
             baodila = " 剩"+baodiNum+"发保底";
@@ -53,10 +59,15 @@ public class PrayListener {
     @Filter(value = "公主十连")
     @Listen(MsgGetTypes.groupMsg)
     public void highTen(GroupMsg msg, MsgSender sender) throws Exception {
-        personService.reduceStar(new Person(msg.getQQ(),TEN,null));
+        String qq = msg.getQQ();
+        CQCode cqCode_at = CQCodeUtil.build().getCQCode_At(qq);
+        if (personService.getStar(qq)<TEN){
+            sender.SENDER.sendGroupMsg(msg.getGroup(),cqCode_at+" 积分不足！");
+            return;
+        }
+        personService.reduceStar(new Person(qq,TEN,null,null));
         String path = prayService.highTen();
         File file = new File(path);
-        CQCode cqCode_at = CQCodeUtil.build().getCQCode_At(msg.getQQ());
         String cqCode_image = CQCodeUtil.build().getCQCode_image("file://" + file.getAbsolutePath());
         sender.SENDER.sendGroupMsg(msg.getGroup(), cqCode_at+cqCode_image);
         file.delete();
@@ -65,10 +76,15 @@ public class PrayListener {
     @Filter(value = "魔女单抽")
     @Listen(MsgGetTypes.groupMsg)
     public void customOne(GroupMsg msg, MsgSender sender) throws Exception {
-        personService.reduceStar(new Person(msg.getQQ(),ONE,null));
-        String high = prayService.customOne(msg.getQQ());
-        CQCode cqCode_at = CQCodeUtil.build().getCQCode_At(msg.getQQ());
-        Integer baodiNum = prayService.getBaodiNum(PrayEnum.custom, msg.getQQ());
+        String qq = msg.getQQ();
+        CQCode cqCode_at = CQCodeUtil.build().getCQCode_At(qq);
+        if (personService.getStar(qq)<ONE){
+            sender.SENDER.sendGroupMsg(msg.getGroup(),cqCode_at+" 积分不足！");
+            return;
+        }
+        personService.reduceStar(new Person(qq,ONE,null,null));
+        String high = prayService.customOne(qq);
+        Integer baodiNum = prayService.getBaodiNum(PrayEnum.custom, qq);
         String baodila = "";
         if (baodiNum<=BAODI){
             baodila = " 剩"+baodiNum+"发保底";
@@ -79,10 +95,59 @@ public class PrayListener {
     @Filter(value = "魔女十连")
     @Listen(MsgGetTypes.groupMsg)
     public void customTen(GroupMsg msg, MsgSender sender) throws Exception {
-        personService.reduceStar(new Person(msg.getQQ(),TEN,null));
+        String qq = msg.getQQ();
+        CQCode cqCode_at = CQCodeUtil.build().getCQCode_At(qq);
+        if (personService.getStar(qq)<TEN){
+            sender.SENDER.sendGroupMsg(msg.getGroup(),cqCode_at+" 积分不足！");
+            return;
+        }
+        personService.reduceStar(new Person(qq,TEN,null,null));
         String path = prayService.customTen();
         File file = new File(path);
-        CQCode cqCode_at = CQCodeUtil.build().getCQCode_At(msg.getQQ());
+        String cqCode_image = CQCodeUtil.build().getCQCode_image("file://" + file.getAbsolutePath());
+        sender.SENDER.sendGroupMsg(msg.getGroup(), cqCode_at.toString()+cqCode_image);
+        file.delete();
+    }
+
+    @Filter(value = "魔法少女单抽")
+    @Listen(MsgGetTypes.groupMsg)
+    public void specialOne(GroupMsg msg, MsgSender sender) throws Exception {
+        String qq = msg.getQQ();
+        CQCode cqCode_at = CQCodeUtil.build().getCQCode_At(qq);
+        if (!PrayUtils.findSpecial()){
+            sender.SENDER.sendGroupMsg(msg.getGroup(), cqCode_at+"魔法少女关门啦！请下次再来！");
+            return;
+        }
+        if (personService.getStar(qq)<TEN){
+            sender.SENDER.sendGroupMsg(msg.getGroup(),cqCode_at+" 积分不足！");
+            return;
+        }
+        personService.reduceStar(new Person(qq,ONE,null,null));
+        String high = prayService.specialOne(qq);
+        Integer baodiNum = prayService.getBaodiNum(PrayEnum.special, qq);
+        String baodila = "";
+        if (baodiNum<=BAODI){
+            baodila = " 剩"+baodiNum+"发保底";
+        }
+        sender.SENDER.sendGroupMsg(msg.getGroup(),cqCode_at+high+baodila);
+    }
+
+    @Filter(value = "魔法少女十连")
+    @Listen(MsgGetTypes.groupMsg)
+    public void specialTen(GroupMsg msg, MsgSender sender) throws Exception {
+        String qq = msg.getQQ();
+        CQCode cqCode_at = CQCodeUtil.build().getCQCode_At(qq);
+        if (!PrayUtils.findSpecial()){
+            sender.SENDER.sendGroupMsg(msg.getGroup(), cqCode_at+"魔法少女关门啦！请下次再来！");
+            return;
+        }
+        if (personService.getStar(qq)<TEN){
+            sender.SENDER.sendGroupMsg(msg.getGroup(),cqCode_at+" 积分不足！");
+            return;
+        }
+        personService.reduceStar(new Person(qq,TEN,null,null));
+        String path = prayService.specialTen();
+        File file = new File(path);
         String cqCode_image = CQCodeUtil.build().getCQCode_image("file://" + file.getAbsolutePath());
         sender.SENDER.sendGroupMsg(msg.getGroup(), cqCode_at.toString()+cqCode_image);
         file.delete();
@@ -93,12 +158,7 @@ public class PrayListener {
     public void middleOne(GroupMsg msg, MsgSender sender) throws Exception {
         String high = prayService.middleOne(msg.getQQ());
         CQCode cqCode_at = CQCodeUtil.build().getCQCode_At(msg.getQQ());
-        Integer baodiNum = prayService.getBaodiNum(PrayEnum.middle, msg.getQQ());
-        String baodila = "";
-        if (baodiNum<=BAODI){
-            baodila = " 剩"+baodiNum+"发保底";
-        }
-        sender.SENDER.sendGroupMsg(msg.getGroup(),cqCode_at+high+baodila);
+        sender.SENDER.sendGroupMsg(msg.getGroup(),cqCode_at+high);
     }
 
     @Filter(value = "大小姐十连")
